fix db tests

+++ internal/api/database/db.go
@@ -2,7 +2,15 @@ package database
 
 import (
 	"database/sql"
+	"errors"
+	"github.com/go-sql-driver/mysql"
 	_ "github.com/go-sql-driver/mysql"
+)
+
+var (
+	ErrDuplicateEntry = errors.New("duplicate entry")
+	ErrNotFound = errors.New("not found")
+	ErrInvalid = errors.New("invalid operation")
 )
 
 type DataBase struct {
@@ -29,7 +37,7 @@ func (db *DataBase) Clear() error {
 func (db *DataBase) AddUser(u User) (int64, error) {
 	res, err := db.db.Exec("INSERT INTO User(Name) VALUES (?)", u.Name)
 	if err != nil {
-		return 0, err
+		return 0, parseError(err)
 	}
 	return res.LastInsertId()
 }
@@ -38,32 +46,61 @@ func (db *DataBase) GetUser(name string) (User, error) {
 	res := db.db.QueryRow("SELECT Id, Name FROM User WHERE Name = ?", name)
 	var u User
 	err := res.Scan(&u.Id, &u.Name)
-	return u, err
+	return u, parseError(err)
 }
 
 func (db *DataBase) DeleteUser(name string) (int64, error) {
 	res, err := db.db.Exec("DELETE FROM User WHERE Name = ?", name)
 	if err != nil {
-		return 0, err
+		return 0, parseError(err)
 	}
 	return res.RowsAffected()
+}
+
+func (db *DataBase) getZoneOwner(zone string) (int64, error) {
+	res := db.db.QueryRow("SELECT User_Id FROM Zone WHERE Name = ?", zone)
+	var userId int64
+	err := res.Scan(&userId)
+	if err != nil {
+		if errors.Is(err, sql.ErrNoRows) {
+			return 0, nil
+		}
+		return -1, err
+	}
+	return userId, nil
 }
 
 func (db *DataBase) AddZone(user string, z Zone) (int64, error) {
 	u, err := db.GetUser(user)
 	if err != nil {
+		if err == ErrNotFound {
+			return 0, ErrInvalid
+		}
 		return 0, err
 	}
-	res, err := db.db.Exec("INSERT INTO Zone(Name, CNameFlattening, Dnssec, Enabled, User_Id) VALUES (?, ?, ?, ?, ?)", z.Name, z.CNameFlattening, z.Dnssec, z.Enabled, u.Id)
+	owner, err := db.getZoneOwner(z.Name)
 	if err != nil {
 		return 0, err
 	}
-	return res.LastInsertId()
+	if owner == 0 {
+		res, err := db.db.Exec("INSERT INTO Zone(Name, CNameFlattening, Dnssec, Enabled, User_Id) VALUES (?, ?, ?, ?, ?)", z.Name, z.CNameFlattening, z.Dnssec, z.Enabled, u.Id)
+		if err != nil {
+			return 0, parseError(err)
+		}
+		return res.LastInsertId()
+	}
+	if owner == u.Id {
+		return 0, ErrDuplicateEntry
+	}
+	return 0, ErrInvalid
 }
 
 func (db *DataBase) GetZones(user string, start int, count int, q string) ([]string, error) {
 	u, err := db.GetUser(user)
 	if err != nil {
+		if err == ErrNotFound {
+			return nil, ErrInvalid
+		}
 		return nil, err
 	}
 	like := "%" + q + "%"
@@ -85,14 +122,14 @@ func (db *DataBase) GetZones(user string, start int, count int, q string) ([]str
 }
 
 func (db *DataBase) GetZone(zone string) (Zone, error) {
-	res := db.db.QueryRow("SELECT Id, Name, Enabled FROM Zone WHERE Name = ?", zone)
+	res := db.db.QueryRow("SELECT Id, Name, CNameFlattening, Dnssec, Enabled FROM Zone WHERE Name = ?", zone)
 	var z Zone
-	err := res.Scan(&z.Id, &z.Name, &z.Enabled)
-	return z, err
+	err := res.Scan(&z.Id, &z.Name, &z.CNameFlattening, &z.Dnssec, &z.Enabled)
+	return z, parseError(err)
 }
 
 func (db *DataBase) UpdateZone(z Zone) (int64, error) {
-	res, err := db.db.Exec("UPDATE Zone SET Enabled = ? WHERE Name = ?", z.Enabled, z.Name)
+	res, err := db.db.Exec("UPDATE Zone SET Dnssec = ?, CNameFlattening = ?, Enabled = ? WHERE Name = ?", z.Dnssec, z.CNameFlattening, z.Enabled, z.Name)
 	if err != nil {
 		return 0, err
 	}
@@ -102,7 +139,7 @@ func (db *DataBase) UpdateZone(z Zone) (int64, error) {
 func (db *DataBase) DeleteZone(zone string) (int64, error) {
 	res, err := db.db.Exec("DELETE FROM Zone WHERE Name = ?", zone)
 	if err != nil {
-		return 0, err
+		return 0, parseError(err)
 	}
 	return res.RowsAffected()
 }
@@ -110,11 +147,14 @@ func (db *DataBase) DeleteZone(zone string) (int64, error) {
 func (db *DataBase) AddLocation(zone string, l Location) (int64, error) {
 	z, err := db.GetZone(zone)
 	if err != nil {
+		if err == ErrNotFound {
+			return 0, ErrInvalid
+		}
 		return 0, err
 	}
 	res, err := db.db.Exec("INSERT INTO Location(Name, Enabled, Zone_Id) VALUES (?, ?, ?)", l.Name, l.Enabled, z.Id)
 	if err != nil {
-		return 0, err
+		return 0, parseError(err)
 	}
 	return res.LastInsertId()
 }
@@ -122,6 +162,9 @@ func (db *DataBase) AddLocation(zone string, l Location) (int64, error) {
 func (db *DataBase) GetLocations(zone string, start int, count int, q string) ([]string, error) {
 	z, err := db.GetZone(zone)
 	if err != nil {
+		if err == ErrNotFound {
+			return nil, ErrInvalid
+		}
 		return nil, err
 	}
 	like := "%" + q + "%"
@@ -145,22 +188,28 @@ func (db *DataBase) GetLocations(zone string, start int, count int, q string) ([
 func (db *DataBase) GetLocation(zone string, location string) (Location, error) {
 	z, err := db.GetZone(zone)
 	if err != nil {
+		if err == ErrNotFound {
+			return Location{}, ErrInvalid
+		}
 		return Location{}, err
 	}
 	res := db.db.QueryRow("SELECT Id, Name, Enabled FROM Location WHERE Zone_Id = ? AND Name = ?", z.Id, location)
 	var l Location
 	err = res.Scan(&l.Id, &l.Name, &l.Enabled)
-	return l, err
+	return l, parseError(err)
 }
 
 func (db *DataBase) UpdateLocation(zone string, l Location) (int64, error) {
 	z, err := db.GetZone(zone)
 	if err != nil {
+		if err == ErrNotFound {
+			return 0, ErrInvalid
+		}
 		return 0, err
 	}
 	res, err := db.db.Exec("UPDATE Location SET Enabled = ? WHERE Zone_Id = ? AND Name = ?", l.Enabled, z.Id, l.Name)
 	if err != nil {
-		return 0, err
+		return 0, parseError(err)
 	}
 	return res.RowsAffected()
 }
@@ -168,6 +217,9 @@ func (db *DataBase) UpdateLocation(zone string, l Location) (int64, error) {
 func (db *DataBase) DeleteLocation(zone string, location string) (int64, error) {
 	z, err := db.GetZone(zone)
 	if err != nil {
+		if err == ErrNotFound {
+			return 0, ErrInvalid
+		}
 		return 0, err
 	}
 	res, err := db.db.Exec("DELETE FROM Location WHERE Zone_Id = ? AND Name = ?", z.Id, location)
@@ -180,11 +232,14 @@ func (db *DataBase) DeleteLocation(zone string, location string) (int64, error) 
 func (db *DataBase) AddRecordSet(zone string, location string, r RecordSet) (int64, error) {
 	l, err := db.GetLocation(zone, location)
 	if err != nil {
+		if err == ErrNotFound {
+			return 0, ErrInvalid
+		}
 		return 0, err
 	}
 	res, err := db.db.Exec("INSERT INTO RecordSet(Location_Id, Type, Value, Enabled) VALUES (?, ?, ?, ?)", l.Id, r.Type, r.Value, r.Enabled)
 	if err != nil {
-		return 0, err
+		return 0, parseError(err)
 	}
 	return res.LastInsertId()
 }
@@ -192,6 +247,9 @@ func (db *DataBase) AddRecordSet(zone string, location string, r RecordSet) (int
 func (db *DataBase) GetRecordSets(zone string, location string) ([]string, error) {
 	l, err := db.GetLocation(zone, location)
 	if err != nil {
+		if err == ErrNotFound {
+			return nil, ErrInvalid
+		}
 		return nil, err
 	}
 	rows, err := db.db.Query("SELECT Type FROM RecordSet WHERE Location_Id = ?", l.Id)
@@ -212,19 +270,31 @@ func (db *DataBase) GetRecordSets(zone string, location string) ([]string, error
 }
 
 func (db *DataBase) GetRecordSet(zone string, location string, rtype string) (RecordSet, error) {
+	if !isRecordSetType(rtype) {
+		return RecordSet{}, ErrInvalid
+	}
 	l, err := db.GetLocation(zone, location)
 	if err != nil {
+		if err == ErrNotFound {
+			return RecordSet{}, ErrInvalid
+		}
 		return RecordSet{}, err
 	}
 	row := db.db.QueryRow("SELECT Id, Type, Value, Enabled FROM RecordSet WHERE Location_Id = ? AND Type = ?", l.Id, rtype)
 	var r RecordSet
 	err = row.Scan(&r.Id, &r.Type, &r.Value, &r.Enabled)
-	return r, err
+	return r, parseError(err)
 }
 
 func (db *DataBase) UpdateRecordSet(zone string, location string, r RecordSet) (int64, error) {
+	if !isRecordSetType(r.Type) {
+		return 0, ErrInvalid
+	}
 	l, err := db.GetLocation(zone , location)
 	if err != nil {
+		if err == ErrNotFound {
+			return 0, ErrInvalid
+		}
 		return 0, err
 	}
 	res, err := db.db.Exec("UPDATE RecordSet SET Value = ?, Enabled = ?  WHERE Location_Id = ? AND Type = ?", r.Value, r.Enabled, l.Id, r.Type)
@@ -235,8 +305,14 @@ func (db *DataBase) UpdateRecordSet(zone string, location string, r RecordSet) (
 }
 
 func (db *DataBase) DeleteRecordSet(zone string, location string, rtype string) (int64, error) {
+	if !isRecordSetType(rtype) {
+		return 0, ErrInvalid
+	}
 	l, err := db.GetLocation(zone, location)
 	if err != nil {
+		if err == ErrNotFound {
+			return 0, ErrInvalid
+		}
 		return 0, err
 	}
 	res, err := db.db.Exec("DELETE FROM RecordSet WHERE Location_Id = ? AND Type = ?", l.Id, rtype)
@@ -244,4 +320,28 @@ func (db *DataBase) DeleteRecordSet(zone string, location string, rtype string) 
 		return 0, err
 	}
 	return res.RowsAffected()
+}
+
+func parseError(err error) error {
+	var mysqlErr *mysql.MySQLError
+	if errors.As(err, &mysqlErr) {
+		if mysqlErr.Number == 1062 {
+			return ErrDuplicateEntry
+		}
+		return err
+	}
+	if errors.Is(err, sql.ErrNoRows) {
+		return ErrNotFound
+	}
+	return err
+}
+
+func isRecordSetType(rtype string) bool {
+	types := []string{"a", "aaaa", "cname", "txt", "ns", "mx", "srv", "caa", "ptr", "tlsa", "ds", "aname"}
+	for _, t := range types {
+		if rtype == t {
+			return true
+		}
+	}
+	return false
 }
